+++ cookbooks/chub_rundeck/recipes/server.rb
@@ -1,10 +1,8 @@
 #
-# Cookbook Name:: Base
-# Recipe:: base_linux-deb
+# Cookbook Name:: chub_rundeck
+# Recipe:: server
 #
-# Cookbook that installs standard packages
-#
-# Copyright 2013, CommerceHub
+# Copyright 2014, CommerceHub Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,37 +17,19 @@
 # limitations under the License.
 #
 
-include_recipe "apt"
+include_recipe "rundeck::default"
 
-package_list = %w{
-	build-essential
-	zsh
-	vim
-	nmap
-	curl
-	wget
-	netcat
-	htop
-	strace
-	sysstat
-	ruby1.9.3
-	mercurial
-}
+directory "/var/lib/rundeck/.ssh" do
+    action :create
+    owner 'rundeck'
+    group 'rundeck'
+    mode 00644
+end 
 
-package_list.each do |pkg|
-	package pkg do
-		action :install
-	end
-end
-
-include_recipe "git"
-include_recipe "chef-kick"
-
-unless node['instance_role'] == 'vagrant'
-	include_recipe "chef-client"
-end
-
-group "minions" do 
-  action :create
-  append true
+file "/var/lib/rundeck/.ssh/id_rsa" do
+    action :create
+    owner 'rundeck-ssh'
+    group 'rundeck-ssh'
+    mode 00644
+    content node[:rundeck][:ssh][:private_key]
 end