+++ cookbooks/base/CHANGELOG.md
@@ -1,5 +1,9 @@
 base CHANGELOG
 ==============
+0.1.15
+----
+- Larry - Adding minions group to all machines by default. 
+
 0.1.14
 ----
 - Larry - Tweaking users recipe to only add users to machine if they are found in ssh_keys has on the node.